cosmos-hub plugin working and added parsing from CLI params
