+++ tools/rosetta/.gitignore
@@ -1,1 +1,3 @@
 /rosetta
+
+*.so
