--- tools/rosetta/plugins/cosmos-hub/Makefile
@@ -1,4 +1,0 @@
-#!/usr/bin/make -f
-
-plugin:
-	go build -buildmode=plugin -o cosmos-hub.so cosmos-hub.go
