+++ tools/rosetta/plugins/osmosis/Makefile
@@ -1,4 +1,4 @@
 #!/usr/bin/make -f
 
 plugin:
-	go build -buildmode=plugin -o osmosis.so osmosis.go
+	go build -buildmode=plugin -o main.so main.go
